% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netlogo_shape.R
\name{get_netlogo_shape}
\alias{get_netlogo_shape}
\title{Download NetLogo shapes from LogoShapes}
\usage{
get_netlogo_shape(
  shape,
  collection = "netlogo-refined",
  dir = tempdir(),
  auth_token = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{shape}{A \code{\link[base:character]{character}} vector indicating the names
of the shapes to download.}

\item{collection}{(optional) A \code{\link[base:character]{character}} string
indicating the collection of shapes to download from
(default: \code{"netlogo-refined"}).}

\item{dir}{(optional) A \code{\link[base:character]{character}} string indicating
the directory where the shapes will be saved (default: \code{tempdir()}).}

\item{auth_token}{(optional) A \code{\link[base:character]{character}} string
indicating a GitHub Personal Access Token (PAT) for authentication
with the GitHub API. This is useful when dealing with rate limits
(default: \code{Sys.getenv("GITHUB_PAT")}).}
}
\value{
A named \code{\link[base:character]{character}} vector with the file paths
to the downloaded NetLogo shapes as
\href{https://en.wikipedia.org/wiki/SVG}{SVG} files.
}
\description{
\code{get_netlogo_shape()} downloads NetLogo shapes from the
\href{https://github.com/danielvartan/logoshapes}{LogoShapes} project on GitHub.

The collections and shapes available for download can be found in the
LogoShapes \href{https://github.com/danielvartan/logoshapes/tree/main/svg}{\code{svg}}
directory. Refer to the
\href{https://github.com/danielvartan/logoshapes}{LogoShapes} documentation for
more information about the different collections.

This function requires an active internet connection and the
\href{https://httr2.r-lib.org/}{\code{httr2}} package.
}
\examples{
\dontrun{
  library(fs)
  library(magick)
}

\dontrun{
  shape <- get_netlogo_shape("turtle")

  file_size(shape)

  shape |> image_read_svg() |> image_ggplot()
}

\dontrun{
  shape <- get_netlogo_shape("turtle", collection = "netlogo-simplified")

  file_size(shape)

  shape |> image_read_svg() |> image_ggplot()
}

\dontrun{
  shape <- get_netlogo_shape("turtle", collection = "netlogo-7-0-3")

  file_size(shape)

  shape |> image_read_svg() |> image_ggplot()
}
}
\seealso{
Other utility functions: 
\code{\link{find_netlogo_console}()},
\code{\link{find_netlogo_home}()},
\code{\link{find_netlogo_version}()},
\code{\link{inspect_experiment_file}()},
\code{\link{parse_netlogo_color}()},
\code{\link{parse_netlogo_list}()}
}
\concept{utility functions}
