% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_experiment_file.R
\name{inspect_experiment_file}
\alias{inspect_experiment_file}
\title{Inspect a BehaviorSpace experiment XML file}
\usage{
inspect_experiment_file(file)
}
\arguments{
\item{file}{A string specifying the path to the BehaviorSpace experiment
XML file.}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\code{inspect_experiment_file()} reads and prints the content of a
BehaviorSpace experiment XML file. This is useful for debugging and
understanding the structure of the experiment file.
}
\examples{
file <- create_experiment(name = "My Experiment")

file |> inspect_experiment_file()
}
\seealso{
Other Utility functions: 
\code{\link{parse_netlogo_list}()}
}
\concept{Utility functions}
