% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logolink-package.R
\docType{package}
\name{logolink-package}
\alias{logolink}
\alias{logolink-package}
\title{logolink: An Interface for Running 'NetLogo' Simulations}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

An interface for 'NetLogo' \url{https://www.netlogo.org} that enables programmatic setup and execution of simulations. Designed to facilitate integrating 'NetLogo' models into reproducible workflows by creating and running 'BehaviorSpace' experiments and retrieving their results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/logolink/}
  \item \url{https://github.com/danielvartan/logolink/}
  \item Report bugs at \url{https://github.com/danielvartan/logolink/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danielvartan@proton.me} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

Authors:
\itemize{
  \item Daniel Vartanian \email{danielvartan@proton.me} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]
}

}
\keyword{internal}
