% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patches.R
\name{plot_patches}
\alias{plot_patches}
\title{Plot NetLogo World Patches}
\usage{
plot_patches(
  results,
  run_number = 1,
  step = 0,
  na_value = parse_netlogo_color(7.5)
)
}
\arguments{
\item{results}{A \code{\link[base:list]{list}} of experiment results from
\code{\link[=run_experiment]{run_experiment()}}. Must contain the \code{lists} element
with the required patch metrics. See the \emph{Details} section for more
information.}

\item{run_number}{(optional) An integer number indicating the specific
simulation run to visualize (default: \code{1}).}

\item{step}{(optional) An integer number indicating the specific time step
within the run to visualize (default: \code{0}).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object representing the NetLogo world patches
at the specified run and step.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{plot_patches()} visualizes the state of NetLogo patches at a specific run
and step. Patches are rendered based on their \code{pcolor} attribute, providing
a spatial snapshot of the environment.

This function serves as a foundational layer for spatial visualization. Users
can extend the resulting plot by adding layers for turtles, links, or other
\href{https://ggplot2.tidyverse.org/}{\code{ggplot2}} components to match their
specific model requirements.
}
\details{
To use this function, your experiment must collect specific patch data.
Ensure the following reporters are included in the \code{metrics} argument of
\code{\link[=run_experiment]{run_experiment()}}:
\itemize{
\item \verb{[pxcor] of patches}
\item \verb{[pycor] of patches}
\item \verb{[pcolor] of patches}
}

These metrics must be available within the \code{lists} output of the results
object.
}
\examples{
# Loading Packages -----

\dontrun{
  library(ggplot2)
  library(ggimage)
  library(magrittr)
  library(stringr)
  library(tidyr)
}

# Defining the Model -----

\dontrun{
  model_path <- # This model is included with NetLogo installations.
    find_netlogo_home() |>
    file.path(
      "models",
      "Sample Models",
      "Biology",
      "Wolf Sheep Predation.nlogox"
    )
}

# Creating an Experiment -----

\dontrun{
  setup_file <- create_experiment(
    name = "Agent Attributes Extraction",
    repetitions = 1,
    sequential_run_order = TRUE,
    run_metrics_every_step = TRUE,
    time_limit = 1,
    metrics = c(
      '[who] of sheep',
      '[xcor] of sheep',
      '[ycor] of sheep',
      '[shape] of sheep',
      '[color] of sheep',
      '[who] of wolves',
      '[xcor] of wolves',
      '[ycor] of wolves',
      '[shape] of wolves',
      '[color] of wolves',
      '[pxcor] of patches',
      '[pycor] of patches',
      '[pcolor] of patches'
    ),
    constants = list(
      "model-version" = "sheep-wolves-grass"
    )
  )
}

# Running the Experiment -----

\dontrun{
  results <-
    model_path |>
    run_experiment(
      setup_file = setup_file,
      outputs = c("table", "lists")
    )
}

# Visualizing Patches -----

\dontrun{
  plot <- results |> plot_patches()

  print(plot)
}

# Getting Agents Shapes -----

\dontrun{
  turtle_shapes <-
    system.file("extdata", package = "logolink") |>
    list.files(full.names = TRUE)
}

# Adding Agents to the Plot -----

\dontrun{
  plot +
    geom_image(
      data = results |>
        extract2("lists") |>
        drop_na(xcor_of_sheep),
      mapping = aes(
        x = xcor_of_sheep,
        y = ycor_of_sheep,
        image = turtle_shapes |>
          str_subset("sheep") |>
          head(1)
      ),
      size = 0.035
    ) +
    geom_image(
      data = results |>
        extract2("lists") |>
        drop_na(xcor_of_wolves),
      mapping = aes(
        x = xcor_of_wolves,
        y = ycor_of_wolves,
        image = turtle_shapes |>
          str_subset("wolf") |>
          head(1)
      ),
      size = 0.04
    )
}
}
\concept{plot functions}
