% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_netlogo_executable.R
\name{find_netlogo_console}
\alias{find_netlogo_console}
\title{Find NetLogo Executable File}
\usage{
find_netlogo_console(netlogo_home = find_netlogo_home(), remove_exe_ext = TRUE)
}
\arguments{
\item{netlogo_home}{(optional) A \code{\link[base:character]{character}} string
specifying the path to the NetLogo installation directory. If not provided,
the function will try to find it automatically using
\code{\link[=find_netlogo_home]{find_netlogo_home()}}. (default:
\code{\link[=find_netlogo_home]{find_netlogo_home()}}).}

\item{remove_exe_ext}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to remove the \code{.exe} extension from the returned path on Windows
systems (default: \code{TRUE}).}
}
\value{
A \code{\link[base:character]{character}} string specifying the path to the
NetLogo executable file. If the file cannot be found, an empty string is
returned.
}
\description{
\code{find_netlogo_console()} attempts to locate the NetLogo executable file on
the user's system.

It first checks the \code{NETLOGO_CONSOLE} environment variable. If this variable
is not set or the file does not exist, it constructs the path to the
executable based on the provided \code{netlogo_home} directory.
}
\examples{
\dontrun{
  find_netlogo_console()
}
}
\seealso{
Other utility functions: 
\code{\link{find_netlogo_home}()},
\code{\link{find_netlogo_version}()},
\code{\link{inspect_experiment_file}()},
\code{\link{parse_netlogo_list}()}
}
\concept{utility functions}
