% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_netlogo_home.R
\name{find_netlogo_home}
\alias{find_netlogo_home}
\title{Find NetLogo Installation Directory}
\usage{
find_netlogo_home()
}
\value{
A \code{\link[base:character]{character}} string specifying the path to the
NetLogo installation directory. If the directory cannot be found,
an empty string is returned.
}
\description{
\code{find_netlogo_home()} attempts to locate the installation directory of
NetLogo on the user's system.

It first checks the \code{NETLOGO_HOME} environment variable. If this variable is
not set or the directory does not exist, it searches through a list of common
installation paths for NetLogo.
}
\examples{
\dontrun{
  find_netlogo_home()
}
}
\seealso{
Other utility functions: 
\code{\link{find_netlogo_console}()},
\code{\link{find_netlogo_version}()},
\code{\link{inspect_experiment_file}()},
\code{\link{parse_netlogo_list}()}
}
\concept{utility functions}
