% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_netlogo_version.R
\name{find_netlogo_version}
\alias{find_netlogo_version}
\title{Find NetLogo Version}
\usage{
find_netlogo_version(netlogo_home = find_netlogo_home())
}
\arguments{
\item{netlogo_home}{(optional) A \code{\link[base:character]{character}} string
specifying the path to the NetLogo installation directory. If not provided,
the function will try to find it automatically using
\code{\link[=find_netlogo_home]{find_netlogo_home()}}. (default:
\code{\link[=find_netlogo_home]{find_netlogo_home()}}).}
}
\value{
A \code{\link[base:character]{character}} string specifying the NetLogo
version installed on the user's system. If the version cannot be
determined, an empty string is returned.
}
\description{
\code{find_netlogo_version()} attempts to locate the NetLogo version installed on
the user's system.

It first tries to execute the NetLogo console with the \code{--version} argument
to retrieve the version information. If the executable is not found, it
attempts to extract the version number from the installation directory name.
}
\examples{
\dontrun{
  find_netlogo_version()
}
}
\seealso{
Other utility functions: 
\code{\link{find_netlogo_console}()},
\code{\link{find_netlogo_home}()},
\code{\link{inspect_experiment_file}()},
\code{\link{parse_netlogo_list}()}
}
\concept{utility functions}
