% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_experiment.R
\name{inspect_experiment}
\alias{inspect_experiment}
\title{Inspect a BehaviorSpace experiment XML file}
\usage{
inspect_experiment(file)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} string specifying the path to
the BehaviorSpace experiment XML file.}
}
\value{
An \link[base:invisible]{invisible} \code{NULL}. This function is called for
its side effect of printing the XML content to the console.
}
\description{
\code{inspect_experiment()} reads and prints the content of a
\href{https://docs.netlogo.org/behaviorspace.html}{BehaviorSpace} experiment XML
file to the console. This is useful for debugging and verifying the
structure of experiment files created by
\code{\link[=create_experiment]{create_experiment()}}.

Please refer to the
\href{https://docs.netlogo.org/behaviorspace.html}{BehaviorSpace Guide} for
complete guidance on how to set and run experiments in NetLogo.
}
\examples{
file <- create_experiment(name = "My Experiment")

file |> inspect_experiment()
}
\seealso{
Other utility functions: 
\code{\link{get_netlogo_shape}()}
}
\concept{utility functions}
