% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_experiment_file.R
\name{inspect_experiment_file}
\alias{inspect_experiment_file}
\title{Inspect a BehaviorSpace experiment XML file}
\usage{
inspect_experiment_file(file)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} string specifying the path to
the BehaviorSpace experiment XML file.}
}
\value{
An \link[base:invisible]{invisible} \code{NULL}. This function is called for
its side effect of printing the XML content to the console.
}
\description{
\code{inspect_experiment_file()} reads and prints the content of a
BehaviorSpace experiment XML file to the console. This is useful for
debugging and verifying the structure of experiment files created by
\code{\link[=create_experiment]{create_experiment()}}.
}
\examples{
file <- create_experiment(name = "My Experiment")

file |> inspect_experiment_file()
}
\seealso{
Other utility functions: 
\code{\link{find_netlogo_console}()},
\code{\link{find_netlogo_home}()},
\code{\link{find_netlogo_version}()},
\code{\link{get_netlogo_shape}()},
\code{\link{parse_netlogo_color}()},
\code{\link{parse_netlogo_list}()}
}
\concept{utility functions}
