% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logolink-package.R
\docType{package}
\name{logolink-package}
\alias{logolink}
\alias{logolink-package}
\title{logolink: Control NetLogo from R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R interface to NetLogo that provides comprehensive tools for programmatically setting up and executing NetLogo agent-based model simulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/logolink/}
  \item \url{https://github.com/danielvartan/logolink/}
  \item Report bugs at \url{https://github.com/danielvartan/logolink/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danielvartan@proton.me} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
