% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run a NetLogo BehaviorSpace experiment}
\usage{
run_experiment(
  netlogo_path,
  model_path,
  experiment = NULL,
  setup_file = NULL,
  other_args = NULL,
  parse = TRUE
)
}
\arguments{
\item{netlogo_path}{A string specifying the path to the NetLogo executable.
In Windows, this is usually something like
\verb{C:/Program Files/NetLogo 7.0.0/NetLogo.exe}.}

\item{model_path}{A string specifying the path to the NetLogo model file
(with extension \code{.nlogo}, \code{.nlogo3d}, \code{.nlogox}, or \code{.nlogox3d}).}

\item{experiment}{(optional) A string specifying the name of the experiment
defined in the NetLogo model file (default: \code{NULL}).}

\item{setup_file}{(optional) A string specifying the path to an XML file
containing the experiment definition. This file can be created using
\code{\link[=create_experiment]{create_experiment()}} or exported from the
NetLogo BehaviorSpace interface (default: \code{NULL}).}

\item{other_args}{(optional) A character vector specifying any additional
command-line arguments to pass to the NetLogo executable. For example,
you can use \code{c("--threads", "4")} to specify the number of threads to use
(default: \code{NULL}).}

\item{parse}{(optional) A boolean indicating whether to parse NetLogo lists
in the output data frame (default: \code{TRUE}). If \code{TRUE}, columns containing
NetLogo lists (e.g., \verb{[1 2 3]}) will be converted to R lists. If \code{FALSE},
the columns will remain as character strings.}
}
\value{
A \code{\link[dplyr:reexports]{tibble}} containing the results of the
experiment.
}
\description{
\code{run_experiment()} runs a NetLogo BehaviorSpace experiment in headless mode
and returns the results as a tidy data frame. It can be used with
\code{\link[=create_experiment]{create_experiment()}} to create the experiment XML
file on the fly, or with an existing experiment defined in an XML file or
stored in the NetLogo model file.

Please refer to the
\href{https://docs.netlogo.org/behaviorspace.html}{BehaviorSpace Guide} for
complete guidance on how to set and run experiments in NetLogo.
}
\examples{
\dontrun{
  1 + 1
}
}
\seealso{
Other NetLogo functions: 
\code{\link{create_experiment}()}
}
\concept{NetLogo functions}
